feat(spans-table,logs-table): do not hide span tags & log fields that has a seperate column
When that seperate column is removed, tags & fields column does not updated. In order to see that
value, you must remove & re-add tags & fields column. The only fix I've discovered is forcefully
redraw the table, which causes the lose scroll position. So I guess it's better to show that columns
even if they have seperate column.